+++ data-raw/alignments_and_sites.R
@@ -27,20 +27,12 @@ hcf_gp120_end <- which(cumsum(as.character(
 ) != "-") == 7757)[1]
 hcf_gp120 <- slice_aln(hxb2_cons_founder, hcf_gp120_start, hcf_gp120_end)
 
-len_hxb2_gp120 <- nchar(
-  extract_seqs(hcf_gp120, "B.FR.83.HXB2_LAI_IIIB_BRU.K03455")$founder
-)
-flt_gp120_start <- 1
-flt_gp120_end <- which(cumsum(as.character(hiv_env_flt_2022_complete[1, ]) !=
-  "-") == len_hxb2_gp120)[1]
-# subset to only gp120 section
-flt_gp120 <- slice_aln(hiv_env_flt_2022_complete, 1, flt_gp120_end)
 
 founder_conserved_sites <-
-  identify_conserved_sites(flt_gp120,
+  identify_conserved_sites(hiv_env_flt_2022_complete,
     founder = "B.US.2011.DEMB11US006.KC473833",
     ref = "B.FR.83.HXB2_LAI_IIIB_BRU.K03455",
-    founder_aln = hcf_gp120
+    founder_aln = hcf_gp120 # keeps only sites in founder_aln
   ) |>
   dplyr::filter(conserved == "Yes") |>
   dplyr::select(founder_pos, consensus_base) |>