+++ inst/python/model/__init__.py
@@ -1,6 +1,6 @@
 from model.seed import set_python_seed
-from model.epitope import Epitope
-from model.variation import get_substitution, get_recomb_breakpoints, get_recombined_sequence, muts_rel_ref, get_conserved_sites_mutated
+from model.epitope import BEpitope, TEpitope
+from model.variation import get_substitution, get_recomb_breakpoints, get_recombined_sequence, muts_rel_ref, get_conserved_sites_mutated, translate
 from model.fitness import calc_seq_fitness, normalize
 from model.virus import HIV
 from model.cells import InfectedCD4