+++ inst/python/model/config.py
@@ -7,7 +7,7 @@ from scipy.linalg import expm
 from math import exp
 
 from model.virus import HIV
-from model.epitope import Epitope
+from model.epitope import BEpitope, TEpitope
 from model.host import HostEnv
 from model.seed import set_python_seed
 
@@ -17,7 +17,8 @@ class Config:
         q, mut_rate, recomb_rate,
         act_to_lat_rate, lat_to_act_rate, lat_die_rate, lat_prolif_rate,
         conserved_sites, conserved_cost, ref_seq, replicative_cost,
-        epitope_locations, immune_start_day, n_for_imm, days_full_potency,
+        b_epitope_locations, b_immune_start_day, b_n_for_imm, b_days_full_potency,
+        t_epitope_locations, t_max_imm, #t_days_full_potency,
         seed
     ):
         self.last_sampled_gen = last_sampled_gen
@@ -40,10 +41,11 @@ class Config:
         self.prob_lat_to_act = self.rate_to_probability(lat_to_act_rate / generation_time)
         self.prob_lat_die = self.rate_to_probability(lat_die_rate / generation_time)
         self.prob_lat_prolif = self.rate_to_probability(lat_prolif_rate / generation_time)
-      
+
         # determine type of recombination breakpoint method to do
-        recrate_is_sparse = True
-        base_prob = self.prob_recomb
+        self.recrate_is_sparse = True
+        self.base_prob = self.prob_recomb
+
         if isinstance(self.prob_recomb, (float, int)):
             self.prob_recomb = np.full(self.seq_len - 1, self.prob_recomb)
         else:
@@ -55,11 +57,49 @@ class Config:
             probs, nprob = np.unique(self.prob_recomb, return_counts=True)
             L = self.seq_len - 1
             maxidx = np.argmax(nprob)
-            base_prob = probs[maxidx]
-            recrate_is_sparse = ((L - nprob[maxidx]) / L) < sparse_threshold
+
+            self.base_prob = probs[maxidx]
+            self.recrate_is_sparse = ((L - nprob[maxidx]) / L) < sparse_threshold
         
-        self.base_prob = base_prob
-        self.recrate_is_sparse = recrate_is_sparse
+        self.b_epitope_locations = b_epitope_locations
+        if b_epitope_locations is not None:
+            self.b_epitope_locations = [BEpitope(int(epi[1]), int(epi[2]), float(epi[3])) for epi in b_epitope_locations.itertuples()]
+        self.b_immune_start_gen = b_immune_start_day / generation_time
+        self.b_n_for_imm = b_n_for_imm
+        # convert days to generations
+        self.b_gen_full_potency = b_days_full_potency / generation_time
+        
+        self.t_epitope_locations = t_epitope_locations
+        self.t_max_imm = t_max_imm
+        self.t_epitope_mask = None
+        self.t_recognition_motifs = None
+        # epitope, start, escape position, recognized_aa
+        if t_epitope_locations is not None:
+            starts = []
+            t_days_to_full_potency = []
+            positions = []
+            aas = []
+            for row in t_epitope_locations.itertuples():
+                starts.append(int(row[1]))
+                t_days_to_full_potency.append(int(row[2]))
+                positions.append(int(row[3]))
+                aas.append(row[4])
+            epitopes = []
+            for start in set(starts):
+                epitopes.append(TEpitope(
+                    start, np.unique(np.array(t_days_to_full_potency)[np.array(starts) == start])[0] / generation_time, np.array(positions)[np.array(starts) == start], np.array(aas)[np.array(starts) == start]))
+            self.t_epitope_locations = epitopes
+
+            t_epitope_mask = np.full(self.seq_len, -1, dtype=int)
+            recognition_motifs = []
+            epi_num = 0
+            for epi in self.t_epitope_locations:
+                for position in epi.escape_positions:
+                    t_epitope_mask[(epi.start + (position-1)*3):(epi.start + (position-1)*3+3)] = epi_num
+                recognition_motifs.append(''.join([epi.escape_positions[pos] if pos in epi.escape_positions else 'N' for pos in range(1, max(epi.escape_positions.keys())+1)]))
+                epi_num += 1
+            self.t_epitope_mask = t_epitope_mask
+            self.t_recognition_motifs = recognition_motifs
         
         self.conserved_sites = {int(k): v.upper() for k, v in conserved_sites.items()}
         self.conserved_cost = conserved_cost
@@ -70,14 +110,6 @@ class Config:
         if(len(ref_seq) > 0):
             self.prep_ref_conserved() # MAKE A TEST TO BE SURE THIS WORKS
         
-        self.epitope_locations = epitope_locations
-        if epitope_locations is not None:
-            self.epitope_locations = [Epitope(int(epi[1]), int(epi[2]), float(epi[3])) for epi in epitope_locations.itertuples()]
-        self.n_for_imm = n_for_imm
-         # convert days to generations
-        self.gen_full_potency = days_full_potency / generation_time
-        self.seroconversion_time = immune_start_day / generation_time
-
         # Get founder viruses
         founder_viruses = [HIV(seq, self) for seq in self.founder_seqs.values()]
         # Create host environment
@@ -122,6 +154,7 @@ class Config:
         for f in founder_virus_sequences:
             diff_sites.update({i for i, (left, right) in enumerate(
                 zip(self.ref_seq, f)) if left != right})
+        # pop these out of conserved if they exist in there
         [self.conserved_sites.pop(x, None) for x in diff_sites]
         # mask conserved sites so they aren't included in replicative fitness computation
         self.ref_seq = "".join(