+++ tests/testthat/test-checks.R
@@ -260,8 +260,25 @@ test_that("check_run_wavess_inputs works", {
     3.5e-5, 1.4e-5,
     NULL, 0.99, NULL, 1,
     NULL, 30, 0.01, 90,
+    NULL, 0.5,
     0.001, 0.01, 0.01, 0.01, NULL
   ))
+  bad_t <- tibble::tibble(
+    start = 0L,
+    days_to_full_potency = 21L,
+    escape_position = 1L
+  )
+  expect_error(
+    check_run_wavess_inputs(
+      inf_pop_size, samp_scheme, fs, 1.2, hiv_q_mat,
+      3.5e-5, 1.4e-5,
+      NULL, 0.99, NULL, 1,
+      NULL, 30, 0.01, 90,
+      bad_t, 0.5,
+      0.001, 0.01, 0.01, 0.01, NULL
+    ),
+    "t_epitope_locations must contain the columns"
+  )
   expect_error(
     check_run_wavess_inputs(
       inf_pop_size, samp_scheme |> dplyr::mutate(n_sample_active = 0), fs, 1.2, hiv_q_mat,
@@ -376,6 +393,7 @@ test_that("check_run_wavess_inputs works", {
     3.5e-5, 1.4e-5,
     c("1" = "a"), 0.99, NULL, 1,
     NULL, 30, 0.01, 90,
+    NULL, 0.5,
     0.001, 0.01, 0.01, 0.01, NULL
   ))
   expect_no_error(check_run_wavess_inputs(
@@ -383,6 +401,7 @@ test_that("check_run_wavess_inputs works", {
     3.5e-5, 1.4e-5,
     c("1" = "A", "2" = "C"), 0.99, NULL, 1,
     NULL, 30, 0.01, 90,
+    NULL, 0.5,
     0.001, 0.01, 0.01, 0.01, NULL
   ))
   expect_error(
@@ -422,6 +441,7 @@ test_that("check_run_wavess_inputs works", {
     3.5e-5, 1.4e-5,
     NULL, 0.99, "ATTT", 0.99,
     NULL, 30, 0.01, 90,
+    NULL, 0.5,
     0.001, 0.01, 0.01, 0.01, NULL
   ))
 
@@ -448,6 +468,7 @@ test_that("check_run_wavess_inputs works", {
     3.5e-5, 1.4e-5,
     NULL, 0.99, NULL, 1,
     el, 30, 0.01, 90,
+    NULL, 0.5,
     0.001, 0.01, 0.01, 0.01, NULL
   ))
   expect_error(
@@ -487,6 +508,7 @@ test_that("check_run_wavess_inputs works", {
       3.5e-5, 1.4e-5,
       NULL, 0.99, NULL, 1,
       el, 30, 10000, 10,
+      NULL, 0.5,
       0.001, 0.01, 0.01, 0.01, NULL
     ),
     "n_for_imm is greater than the maximum population size so there will be"
@@ -497,6 +519,7 @@ test_that("check_run_wavess_inputs works", {
     3.5e-5, 1.4e-5,
     NULL, 0.99, NULL, 1,
     NULL, 30, 0.01, 90,
+    NULL, 0.5,
     0.001, 0.01, 0.01, 0.01, 1
   ))
   expect_error(
@@ -505,6 +528,7 @@ test_that("check_run_wavess_inputs works", {
       3.5e-5, 1.4e-5,
       NULL, 0.99, NULL, 1,
       NULL, 30, 0.01, 90,
+      NULL, 0.5,
       0.001, 0.01, 0.01, 0.01, "a"
     ),
     "seed must be numeric, but is a character"